Clean up verbose and redundant code constructs

+++ app/src/main/java/com/caverock/androidsvg/CSSParser.java
@@ -999,11 +999,11 @@ class CSSParser
       private int  hexChar(int ch)
       {
          if (ch >= '0' && ch <= '9')
-            return ((int)ch - (int)'0');
+            return (ch - (int)'0');
          if (ch >= 'A' && ch <= 'F')
-            return ((int)ch - (int)'A') + 10;
+            return (ch - (int)'A') + 10;
          if (ch >= 'a' && ch <= 'f')
-            return ((int)ch - (int)'a') + 10;
+            return (ch - (int)'a') + 10;
          return -1;
       }
